+++ src/LB_comm/shmem_procinfo.h
@@ -36,8 +36,13 @@ int shmem_procinfo_ext__finalize(void);
 int shmem_procinfo_ext__postfinalize(pid_t pid, int return_stolen);
 
 /* Get / Set Process mask */
-int shmem_procinfo__getprocessmask(pid_t pid, cpu_set_t *mask);
-int shmem_procinfo__setprocessmask(pid_t pid, const cpu_set_t *mask);
+typedef enum QueryType {
+    QUERY_SYNC,
+    QUERY_ASYNC
+} query_t;
+
+int shmem_procinfo__getprocessmask(pid_t pid, cpu_set_t *mask, query_t query);
+int shmem_procinfo__setprocessmask(pid_t pid, const cpu_set_t *mask, query_t query);
 
 /* Generic Getters */
 int  shmem_procinfo__polldrom(pid_t pid, int *new_cpus, cpu_set_t *new_mask);